ai-server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
google-genai
chromadb
tree-sitter
GitPython
python-multipart
structlog
pytest
httpx
sentence-transformers
